FIX: move check function to avoid circular import
