+++ mapie/utils.py
@@ -16,7 +16,6 @@ from sklearn.utils.validation import (_check_sample_weight, _num_features,
 
 from ._compatibility import np_quantile
 from ._typing import ArrayLike, NDArray
-from .conformity_scores import AbsoluteConformityScore, ConformityScore
 
 SPLIT_STRATEGIES = ["uniform", "quantile", "array split"]
 
@@ -597,40 +596,6 @@ def check_lower_upper_bounds(
     if any_inversion:
         warnings.warn(
             "WARNING: The predictions are ill-sorted."
-        )
-
-
-def check_conformity_score(
-    conformity_score: Optional[ConformityScore],
-    sym: bool = True,
-) -> ConformityScore:
-    """
-    Check parameter ``conformity_score``.
-
-    Raises
-    ------
-    ValueError
-        If parameter is not valid.
-
-    Examples
-    --------
-    >>> from mapie.utils import check_conformity_score
-    >>> try:
-    ...     check_conformity_score(1)
-    ... except Exception as exception:
-    ...     print(exception)
-    ...
-    Invalid conformity_score argument.
-    Must be None or a ConformityScore instance.
-    """
-    if conformity_score is None:
-        return AbsoluteConformityScore(sym=sym)
-    elif isinstance(conformity_score, ConformityScore):
-        return conformity_score
-    else:
-        raise ValueError(
-            "Invalid conformity_score argument.\n"
-            "Must be None or a ConformityScore instance."
         )
